requirements.txt
blinker==1.6.3
build==1.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.11.0
elasticsearch==8.11.1
filelock==3.13.1
Flask==3.0.0
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pip-tools==7.3.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
